Cache move to BaseHandler
