+++ system/Cache/Handlers/MemcachedHandler.php
@@ -11,10 +11,8 @@
 
 namespace CodeIgniter\Cache\Handlers;
 
-use CodeIgniter\Cache\CacheInterface;
 use CodeIgniter\Exceptions\CriticalError;
 use Config\Cache;
-use Closure;
 use Exception;
 use Memcache;
 use Memcached;
@@ -22,7 +20,7 @@ use Memcached;
 /**
  * Mamcached cache handler
  */
-class MemcachedHandler implements CacheInterface
+class MemcachedHandler extends BaseHandler
 {
 	/**
 	 * Prefixed to all cache names.
@@ -193,31 +191,6 @@ class MemcachedHandler implements CacheInterface
 		}
 
 		return is_array($data) ? $data[0] : $data; // @phpstan-ignore-line
-	}
-
-	//--------------------------------------------------------------------
-
-	/**
-	 * Get an item from the cache, or execute the given Closure and store the result.
-	 *
-	 * @param string  $key      Cache item name
-	 * @param integer $ttl      Time to live
-	 * @param Closure $callback Callback return value
-	 *
-	 * @return mixed
-	 */
-	public function remember(string $key, int $ttl, Closure $callback)
-	{
-		$value = $this->get($key);
-
-		if (! is_null($value))
-		{
-			return $value;
-		}
-
-		$this->save($key, $value = $callback(), $ttl);
-
-		return $value;
 	}
 
 	//--------------------------------------------------------------------
