+++ system/Cache/Handlers/RedisHandler.php
@@ -11,17 +11,15 @@
 
 namespace CodeIgniter\Cache\Handlers;
 
-use CodeIgniter\Cache\CacheInterface;
 use CodeIgniter\Exceptions\CriticalError;
 use Config\Cache;
-use Closure;
 use Redis;
 use RedisException;
 
 /**
  * Redis cache handler
  */
-class RedisHandler implements CacheInterface
+class RedisHandler extends BaseHandler
 {
 	/**
 	 * Prefixed to all cache names.
@@ -160,31 +158,6 @@ class RedisHandler implements CacheInterface
 			default:
 				return null;
 		}
-	}
-
-	//--------------------------------------------------------------------
-
-	/**
-	 * Get an item from the cache, or execute the given Closure and store the result.
-	 *
-	 * @param string  $key      Cache item name
-	 * @param integer $ttl      Time to live
-	 * @param Closure $callback Callback return value
-	 *
-	 * @return mixed
-	 */
-	public function remember(string $key, int $ttl, Closure $callback)
-	{
-		$value = $this->get($key);
-
-		if (! is_null($value))
-		{
-			return $value;
-		}
-
-		$this->save($key, $value = $callback(), $ttl);
-
-		return $value;
 	}
 
 	//--------------------------------------------------------------------
