+++ system/Cache/Handlers/WincacheHandler.php
@@ -11,16 +11,14 @@
 
 namespace CodeIgniter\Cache\Handlers;
 
-use CodeIgniter\Cache\CacheInterface;
 use Config\Cache;
-use Closure;
 
 /**
  * Cache handler for WinCache from Microsoft & IIS.
  * Windows-only, so not testable on travis-ci.
  * Unusable methods flagged for code coverage ignoring.
  */
-class WincacheHandler implements CacheInterface
+class WincacheHandler extends BaseHandler
 {
 	/**
 	 * Prefixed to all cache names.
@@ -73,31 +71,6 @@ class WincacheHandler implements CacheInterface
 
 		// Success returned by reference from wincache_ucache_get()
 		return ($success) ? $data : null;
-	}
-
-	//--------------------------------------------------------------------
-
-	/**
-	 * Get an item from the cache, or execute the given Closure and store the result.
-	 *
-	 * @param string  $key      Cache item name
-	 * @param integer $ttl      Time to live
-	 * @param Closure $callback Callback return value
-	 *
-	 * @return mixed
-	 */
-	public function remember(string $key, int $ttl, Closure $callback)
-	{
-		$value = $this->get($key);
-
-		if (! is_null($value))
-		{
-			return $value;
-		}
-
-		$this->save($key, $value = $callback(), $ttl);
-
-		return $value;
 	}
 
 	//--------------------------------------------------------------------
